+++ apprun-v0.3.0/src/store.ts
@@ -69,4 +69,5 @@ export default class Store {
             this.data[9] = a;
         }
     }
+    
 }
