+++ index.html
@@ -18,7 +18,7 @@
             <ul class="nav nav-pills nav-stacked">
                 <li><a href="angular-v1.5.8/index.html"><span class="glyphicon glyphicon-arrow-right" aria-hidden="true"></span> Angular v1.5.8</a></li>
                 <li><a href="angular-v2.0.0/index.html"><span class="glyphicon glyphicon-arrow-right" aria-hidden="true"></span> Angular v2.0.0</a></li>
-                <li><a href="apprun-v0.2.0/index.html"><span class="glyphicon glyphicon-arrow-right" aria-hidden="true"></span> AppRun v0.2.0</a></li>
+                <li><a href="apprun-v0.3.0/index.html"><span class="glyphicon glyphicon-arrow-right" aria-hidden="true"></span> AppRun v0.3.0</a></li>
                 <li><a href="aurelia-v1.0.0/dist/index.html"><span class="glyphicon glyphicon-arrow-right" aria-hidden="true"></span> Aurelia v1.0.0</a></li>
                 <li><a href="bobril-v4.44.1/index.html"><span class="glyphicon glyphicon-arrow-right" aria-hidden="true"></span> Bobril 4.44.1</a></li>
                 <li><a href="cyclejs-v7.0.0/index.html"><span class="glyphicon glyphicon-arrow-right" aria-hidden="true"></span> Cyclejs v7.0.0</a></li>
